arm64: defconfig: add some Kconfig entries used by Tegra132 Norrin FFD board
Add some Kconfig entries to the ARM64 defconfig for functionality used
by the Tegra132, or Tegra K1 64-bit, SoC on the Norrin FFD board.

Once the required compilation fixes to support ARCH_TEGRA builds on
ARM64 have been merged, a subsequent patch will be sent to remove the
CONFIG_BROKEN dependency and to add a few more Kconfig entries that
have ARCH_TEGRA dependencies.  The objective of this patch is to start
building some of the needed code as early as possible to prevent
regressions and to facilitate testing.

This patch draws on work done by Allen Martin <amartin@nvidia.com> and
Thierry Reding <treding@nvidia.com>.

Signed-off-by: Paul Walmsley <paul@pwsan.com>
Cc: Paul Walmsley <pwalmsley@nvidia.com>
Cc: Allen Martin <amartin@nvidia.com>
Cc: Thierry Reding <treding@nvidia.com>
Cc: Catalin Marinas <catalin.marinas@arm.com>
Cc: Will Deacon <will.deacon@arm.com>
Cc: Mark Rutland <mark.rutland@arm.com>